Move sled_agent::opte to illumos_utils::opte
