+++ illumos-utils/Cargo.toml
@@ -11,6 +11,8 @@ cfg-if.workspace = true
 ipnetwork.workspace = true
 macaddr.workspace = true
 omicron-common.workspace = true
+oxide-vpc.workspace = true
+schemars.workspace = true
 serde.workspace = true
 slog.workspace = true
 smf.workspace = true
