+++ illumos-utils/src/lib.rs
@@ -10,6 +10,7 @@ pub mod addrobj;
 pub mod dladm;
 pub mod fstyp;
 pub mod link;
+pub mod opte;
 pub mod running_zone;
 pub mod svc;
 pub mod zfs;
