+++ illumos-utils/src/opte/illumos/port.rs
@@ -4,11 +4,11 @@
 
 //! A single port on the OPTE virtual switch.
 
+use crate::dladm::Dladm;
 use crate::opte::BoundaryServices;
 use crate::opte::Gateway;
 use crate::opte::Vni;
-use crate::params::SourceNatConfig;
-use illumos_utils::dladm::Dladm;
+use crate::opte::params::SourceNatConfig;
 use ipnetwork::IpNetwork;
 use macaddr::MacAddr6;
 use std::net::IpAddr;
@@ -156,7 +156,7 @@ impl Port {
     }
 }
 
-impl illumos_utils::running_zone::OptePort for Port {
+impl crate::running_zone::OptePort for Port {
     fn vnic_name(&self) -> &str {
         Port::vnic_name(self)
     }
