+++ illumos-utils/src/opte/illumos/port_manager.rs
@@ -4,18 +4,18 @@
 
 //! Manager for all OPTE ports on a Helios system
 
+use crate::dladm::Dladm;
+use crate::dladm::PhysicalLink;
+use crate::dladm::VnicSource;
 use crate::opte::default_boundary_services;
 use crate::opte::opte_firewall_rules;
 use crate::opte::Error;
 use crate::opte::Gateway;
 use crate::opte::Port;
 use crate::opte::Vni;
-use crate::params::NetworkInterface;
-use crate::params::SourceNatConfig;
-use crate::params::VpcFirewallRule;
-use illumos_utils::dladm::Dladm;
-use illumos_utils::dladm::PhysicalLink;
-use illumos_utils::dladm::VnicSource;
+use crate::opte::params::NetworkInterface;
+use crate::opte::params::SourceNatConfig;
+use crate::opte::params::VpcFirewallRule;
 use ipnetwork::IpNetwork;
 use macaddr::MacAddr6;
 use opte_ioctl::OpteHdl;
@@ -148,14 +148,10 @@ impl PortManager {
     /// interfaces
     pub fn new(
         log: Logger,
+        data_link: PhysicalLink,
         underlay_ip: Ipv6Addr,
         gateway_mac: MacAddr6,
     ) -> Self {
-        let data_link = crate::common::underlay::find_chelsio_links()
-            .unwrap()
-            .into_iter()
-            .next()
-            .unwrap();
         let inner = Arc::new(PortManagerInner {
             log,
             next_port_id: AtomicU64::new(0),
